Revert "[GCS] refactor the GCS Client Task Interface (#5515)" (#6543)
This reverts commit f78583147c0df1bdec7f1e50750fae8b1993a74e.
