+++ src/ray/core_worker/actor_manager.cc
@@ -1,5 +1,5 @@
 #include "ray/core_worker/actor_manager.h"
-#include "ray/gcs/redis_accessor.h"
+#include "ray/gcs/redis_actor_info_accessor.h"
 
 namespace ray {
