+++ src/ray/gcs/redis_gcs_client.h
@@ -18,14 +18,12 @@ namespace gcs {
 class RedisContext;
 
 class RAY_EXPORT RedisGcsClient : public GcsClient {
-  // TODO(micafan) Will remove those friend class / method after we replace RedisGcsClient
+  // TODO(micafan) Will remove those friend class after we replace RedisGcsClient
   // with interface class GcsClient in raylet.
   friend class RedisActorInfoAccessor;
   friend class RedisJobInfoAccessor;
-  friend class RedisTaskInfoAccessor;
   friend class SubscriptionExecutorTest;
   friend class LogSubscribeTestHelper;
-  friend class TaskTableTestHelper;
 
  public:
   /// Constructor of RedisGcsClient.
@@ -57,6 +55,7 @@ class RAY_EXPORT RedisGcsClient : public GcsClient {
 
   // TODO: Some API for getting the error on the driver
   ObjectTable &object_table();
+  raylet::TaskTable &raylet_task_table();
   TaskReconstructionLog &task_reconstruction_log();
   TaskLeaseTable &task_lease_table();
   ClientTable &client_table();
@@ -97,8 +96,6 @@ class RAY_EXPORT RedisGcsClient : public GcsClient {
   ActorTable &actor_table();
   /// This method will be deprecated, use method Jobs() instead.
   JobTable &job_table();
-  /// This method will be deprecated, use method Tasks() instead.
-  raylet::TaskTable &raylet_task_table();
 
   // GCS command type. If CommandType::kChain, chain-replicated versions of the tables
   // might be used, if available.
