+++ src/ray/gcs/subscription_executor.cc
@@ -189,7 +189,6 @@ Status SubscriptionExecutor<ID, Data, Table>::AsyncUnsubscribe(
 template class SubscriptionExecutor<ActorID, ActorTableData, ActorTable>;
 template class SubscriptionExecutor<ActorID, ActorTableData, DirectActorTable>;
 template class SubscriptionExecutor<JobID, JobTableData, JobTable>;
-template class SubscriptionExecutor<TaskID, TaskTableData, raylet::TaskTable>;
 
 }  // namespace gcs
