+++ src/ray/gcs/test/accessor_test_base.h
@@ -7,7 +7,6 @@
 #include <thread>
 #include <vector>
 #include "gtest/gtest.h"
-#include "ray/gcs/redis_accessor.h"
 #include "ray/gcs/redis_gcs_client.h"
 #include "ray/util/test_util.h"
